Update useSuperHeroData hook to use initialData
###Explanation:
The diff shows that the useSuperHeroData hook is being updated to use the initialData option of the useQuery hook. This option allows you to specify a function that will be called to fetch the initial data for the query. In this case, the function is used to fetch the superhero data from the local query cache. If the data is not found in the cache, the function will fetch it from the server.